Moved map key parameterized tests into their own file
